+++ porespy/tools/__init__.py
@@ -71,33 +71,33 @@ __all__ = [
     "get_tqdm",
     "show_docstring"]
 
-from .__funcs__ import align_image_with_openpnm
-from .__funcs__ import bbox_to_slices
-from .__funcs__ import extend_slice
-from .__funcs__ import extract_cylinder
-from .__funcs__ import extract_subsection
-from .__funcs__ import extract_regions
-from .__funcs__ import find_outer_region
-from .__funcs__ import get_border
-from .__funcs__ import get_planes
-from .__funcs__ import insert_cylinder
-from .__funcs__ import insert_sphere
-from .__funcs__ import in_hull
-from .__funcs__ import isolate_object
-from .__funcs__ import marching_map
-from .__funcs__ import make_contiguous
-from .__funcs__ import mesh_region
-from .__funcs__ import norm_to_uniform
-from .__funcs__ import overlay
-from .__funcs__ import randomize_colors
-from .__funcs__ import recombine
-from .__funcs__ import ps_ball
-from .__funcs__ import ps_disk
-from .__funcs__ import ps_rect
-from .__funcs__ import ps_round
-from .__funcs__ import subdivide
-from .__funcs__ import sanitize_filename
-from .__utils__ import get_tqdm
-from .__utils__ import show_docstring
-from .__funcs__ import _check_for_singleton_axes
+from ._funcs import align_image_with_openpnm
+from ._funcs import bbox_to_slices
+from ._funcs import extend_slice
+from ._funcs import extract_cylinder
+from ._funcs import extract_subsection
+from ._funcs import extract_regions
+from ._funcs import find_outer_region
+from ._funcs import get_border
+from ._funcs import get_planes
+from ._funcs import insert_cylinder
+from ._funcs import insert_sphere
+from ._funcs import in_hull
+from ._funcs import isolate_object
+from ._funcs import marching_map
+from ._funcs import make_contiguous
+from ._funcs import mesh_region
+from ._funcs import norm_to_uniform
+from ._funcs import overlay
+from ._funcs import randomize_colors
+from ._funcs import recombine
+from ._funcs import ps_ball
+from ._funcs import ps_disk
+from ._funcs import ps_rect
+from ._funcs import ps_round
+from ._funcs import subdivide
+from ._funcs import sanitize_filename
+from ._utils import get_tqdm
+from ._utils import show_docstring
+from ._funcs import _check_for_singleton_axes
 from ._unpadfunc import unpad
