+++ porespy/tools/_funcs.py
